Reuse timestamp link in CommentBlock
