+++ packages/web/src/components/comments/CommentText.tsx
@@ -1,72 +1,16 @@
 import { useEffect, useRef, useState } from 'react'
 
 import { useCurrentCommentSection } from '@audius/common/context'
-import { useGatedContentAccess } from '@audius/common/hooks'
 import { commentsMessages as messages } from '@audius/common/messages'
-import { ModalSource } from '@audius/common/models'
-import {
-  playerActions,
-  PurchaseableContentType,
-  trackPageLineupActions,
-  trackPageSelectors,
-  usePremiumContentPurchaseModal
-} from '@audius/common/store'
 import {
   getDurationFromTimestampMatch,
   timestampRegex
 } from '@audius/common/utils'
 import { Flex, Text, TextLink } from '@audius/harmony'
-import { useDispatch, useSelector } from 'react-redux'
 
 import { UserGeneratedTextV2 } from 'components/user-generated-text/UserGeneratedTextV2'
 
-const { tracksActions } = trackPageLineupActions
-const { getLineup } = trackPageSelectors
-const { seek } = playerActions
-
-type TimestampLinkProps = {
-  timestamp: string
-  timestampSeconds: number
-}
-
-const TimestampLink = (props: TimestampLinkProps) => {
-  const { timestamp, timestampSeconds } = props
-
-  const dispatch = useDispatch()
-  const { track } = useCurrentCommentSection()
-  const lineup = useSelector(getLineup)
-  const { track_id: trackId } = track
-
-  const { hasStreamAccess } = useGatedContentAccess(track)
-
-  const uid = lineup?.entries?.[0]?.uid
-  const { onOpen: openPremiumContentPurchaseModal } =
-    usePremiumContentPurchaseModal()
-
-  return (
-    <TextLink
-      onClick={() => {
-        if (!hasStreamAccess) {
-          openPremiumContentPurchaseModal(
-            { contentId: trackId, contentType: PurchaseableContentType.TRACK },
-            {
-              source: ModalSource.Comment
-            }
-          )
-        } else {
-          dispatch(tracksActions.play(uid))
-          setTimeout(() => {
-            dispatch(seek({ seconds: timestampSeconds }))
-          })
-        }
-      }}
-      variant='visible'
-      size='s'
-    >
-      {timestamp}
-    </TextLink>
-  )
-}
+import { TimestampLink } from './TimestampLink'
 
 export type CommentTextProps = {
   children: string
@@ -124,7 +68,6 @@ export const CommentText = (props: CommentTextProps) => {
                 return showLink ? (
                   <TimestampLink
                     key={index}
-                    timestamp={text}
                     timestampSeconds={timestampSeconds}
                   />
                 ) : null
